academicClass controller added

+++ StudentRegistration/Models/AcademicClass.cs
@@ -7,7 +7,7 @@ namespace StudentRegistration.Models
 {
     public class AcademicClass
     {
-        public Int64 ID { get; set; }
+        public int ID { get; set; }
         public string Name { get; set; }
         public ICollection<Student> Students { get; set; }
